UC20 Frontend, UC18 Frontend, Persistence design change and refactor
UC20 Frontend:
-You can now enter vaccine office visits on the office visits page for
HCP
-You should only see appointments that are within 3 hours of the current
time
-There is no error checking yet, this will be added later

UC18 Frontend:
-Added vaccine field to the form for submitting a request
-Added better error checking for selecting a valid date and time
(previously it wouldn't let you select a date & time on the same day, I
changed it so that you can choose the same date so long as its a time in
the future). This was needed to test the 3 hour rule for vaccine
appointments
-I did all of this so that I could enter in some appointment requests so
I could test my front end for UC20

Persistence reworks:
-I got rid of the VaccineDose class and the functionality that went
along with it in the Patient/PatientForm class. This functionality was
causing editing the patient demographics to stop working. I changed it
to search through the database for any previous vaccine appointments for
the patient and then count up the doses for the patient that way (kinda
like Kai suggested in class). All JUnit tests are still passing.
-I changed the VaccineAppointmentRequest field in VaccineOfficeFisitForm
to be the type of Long instead of VaccineAppointmentRequest because the
form classes are used to deserialize JSON data and Spring would have a
hard time deserializeing a custom object like VaccineAppointmentReqeust